Restructuring submodules and removing redundancy
VCTDatabase will hold the current database. getDB can call this. started getting VCTExtraction caught up with this (I think it is now). but it's all a little cludgier than I'd like. I'll probably consider turning VCTConfiguration and VCTDatabase into just files (not modules) and include them in VCTModule which will be the actual module. VCTExtraction would probably have the same fate? I really don't see a reason to be worried about this getting too heavy since I'm going to be running lots of expensive PhysiCell sims and so the loading cost is negligible.