Samplings up and running
Also not allowing for duplicate monads or samplings. This means that when a monad is created, all other monads are searched first to see if they match the current tuple (base_config_id,ic_id,custom_code_id,variation_id). If they do, use the previous monad id that matches. Similarly for samplings, if the monad ids (grabbed from previous) matches (as a set) the monad ids of another sampling, then just use that sampling