Make copies of config files, cut back on global variables
Within config/, make a directory for each base_config_file and populate with all the variation config files. These will be temporary and deleted later.
This then justifies removing current_base_config_id and current_variation_id from the global scope since once the copy is made, simulations just use it.

Also, make xml_doc something to be read in each time. That is what is happening all the time anyways, so having it as a global was not saving any computing resources.